Move EntitiesTable to top-level component
