+++ CHANGELOG-entities-table.md
@@ -1,0 +1,1 @@
+- Refactor entities table to make it reusable.
