+++ context/app/static/js/components/EntitiesTable/style.js
