+++ context/app/static/js/components/EntitiesTable/utils.js
